requirements.txt
connect-openapi-client>=22.0.5
